+++ kameleoon/client/src/Configuration/Rule.php
@@ -2,7 +2,9 @@
 
 namespace Kameleoon\Configuration;
 
-class Rule extends TargetingObject
+use Kameleoon\Targeting\TargetingSegment;
+
+class Rule
 {
     public const EXPERIMENTATION = "EXPERIMENTATION";
     public const TARGETED_DELIVERY = "TARGETED_DELIVERY";
@@ -13,16 +15,18 @@ class Rule extends TargetingObject
     public Experiment $experiment;
     public $exposition;
     public ?int $respoolTime;
+    public ?TargetingSegment $segment;
 
     public function __construct($rule, array $segments)
     {
-        parent::__construct($rule, $segments);
         $this->id = $rule->id ?? 0;
         $this->order = $rule->order;
         $this->type = $rule->type;
         $this->experiment = new Experiment($rule);
         $this->exposition = $rule->exposition;
         $this->respoolTime = $rule->respoolTime;
+        $segmentId = $rule->segmentId ?? null;
+        $this->segment = ($segmentId !== null) ? $segments[$segmentId] ?? null : null;
     }
 
     public function isTargetedDelivery(): bool