+++ kameleoon/client/src/Targeting/TargetingManager.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Kameleoon\Targeting;
 
-use Kameleoon\Configuration\TargetingObject;
-
 interface TargetingManager
 {
-    public function checkTargeting(string $visitorCode, int $containerID, TargetingObject $targetingObject): bool;
+    public function checkTargeting(string $visitorCode, ?int $containerID, ?TargetingSegment $segment): bool;
 }